Add files via upload
It is a simple program, that contains an input and output, a program that creates and uses a list of information. Make sure the program includes a Function that has a PARAMETER within that function a FOR statement that iterates through a List and an IF/ELSE statement. He has ideas but not the knowledge to program them. Ideas: a list of the top 25 universities in The USA with their GPA and SAT requirements. You input your GPA and SAT, and it outputs the Universities that you should select. If you do not have the requirements for both, says a message