Add files via upload
Project 3
In this project, we implement the steepest descent method (SD) and the BFGS method for nding
a local minimizer of a multivariate function. We consider a two dimensional quadratic function
f(x) = 100(x1 􀀀 x2)2 + (x2 􀀀 1)2; (1)
which has a unique minimizer at x? =
"
1
1
#
:
1. Given xk and the search direction pk, the ideal step size k is determined by minimizing the
1-dimensional function '() = f(xk + kpk). For the quadratic function f dened in (1), nd
the explicit formula for k.
2. In general, there is no analytical formula for the best k, nor is it necessary. Implement the
Backtracking algorithm (in note Lec23) for nding an acceptable stepsize (and update xk+1),
using  = 0:001 and  = 0:9.
3. For the quadratic function f given in (1), implement the SD algorithm with backtracking for k
for nding the local minimizer. Take the error tolerance to be 10􀀀10 and x0 = 0, output the error
jjxk 􀀀 x?jj and the iteration number k into a data le. Plot the error vs iteration number.
4. Likewise, implement the BFGS algorithm (note Lec25). Repeat the same test, and plot error vs
iteration number.
Please attach your code to the report in the PDF format.