Add files via upload
DNS Client
This project is a DNS client implemented in Python that can query a DNS server for domain name to IP address translation. It allows the user to provide a hostname as a command-line argument and retrieves the corresponding IP addresses using the DNS protocol.

Features
Reads a hostname provided as a command-line argument.
Prepares a DNS query message adhering to the DNS protocol specifications.
Sends the DNS query message to a DNS server using UDP socket communication.
Receives and processes the DNS server's response.
Extracts and displays the IP addresses from the response.