PyChess application
Online Gameplay consists of two elements, client and server.

CLIENT: It is the machine that runs My-PyChess code.
SERVER: It is a machine that accepts connections from different clients(upto 10)

Back when I was actively working on this project, the networking element of this
app was made with security in mind. I had not used the python "pickle" module,
which is normally used in such applications. This has security risks associated.

The server keeps absolutely NO DATA of any user that connects to it.
No names, accounts, emails, IP address, Location - nothing. PRIVACY IS RESPECTED.

The client and server speak over Raw TCP protocol with no sort of encryption.
This means that any bad guys may read the messages that the client and server
send to one another. But the best part is the fact that the client and server DO NOT
EXCHANGE ANY SENSITIVE DATA. This means that the information is practically useless
for the bad guys. So they will not invest their time and efforts to do such things.
If they did so, they would be in major dissapointment.

The public global test server is not being hosted anymore.